fuzzer: Track the rate of fuzzing in cycles/s.
As libfuzzer develops or as users design their own fuzz tests which use
libfuzzer, it will be essential to understand the performance impact of
changes. This change tracks the rate of execution in terms of cycles per
second and updates the UI at most every second.

This is implemented with a periodic update to the web UI which occurs at
least every 1 second. The period is not guaranteed to be regular.