Use clearer wording and describe `break`
I'm also moving away from "statement" here. Break is actually an
expression, and while it's almost impossible to observe this in code
that doesn't have any compiler errors, it can be shown:

    test "a" {
        _ = while (true) {
            _ = break 42;
        };
    }

Nevertheless, `break` is still described as a statement in the Zig
docs. It may be that the Zig docs simply take statement to mean
something else---it says, "compile log statements", later in the
document.