audit isPowerOfTwo in Sema
The only suspicious part was the isPowerOfTwo in zirShl and zirShr.
This code should trigger a crash in zirShl because the values are not
comptime-known so it's going to reach the bottom where it checks whether
safety is needed and then does isPowerOfTwo() but it doesn't crash:
```
extern const a: u8;

export fn x() void {
    var idk: u0 = @truncate(a);
    _ = &idk;

    const y: u0 = @truncate(a);
    _ = y << idk;
}
```
I think this is because of OVP. So it's not going to reach it anyway.

The only thing I found is some error messages being wrong. This fixes
that.