Linux: Reimplement *stat via statx
Modern Linux ports do not implement the classic `stat` family, instead
relying on libc to implement it via `statx`. Since the minimum kernel we
support is 4.19 we can rely on the presence of `statx`, since it was
released in version 4.11.

The logic has been ported from musl, along with marking parameters as
`noalias`. Also, the `makedev`, `major` and `minor` functions have been
added.