Linux: Introduce generic 64-bit Stat structure, revert statx impl.
Mapping statx -> kernel stat is a fools errand - on modern 32-bit
targets like riscv32 the `stat64` struct uses 32-bit time!
Instead, do what the libc's do and use a generic `stat` structure and
map the arch-specific `stat`s and `statx` onto it.

To help aid this, a compile error has been added to `fstatat` if the
architecture does not support it.