Add implementation for base32 to the stdlib
This is an adaptation of the base64 implementation already in the stdlib. Before
I go further with this I want to test the waters a little. I couldn't find any
issues related to this, so I'm unsure whether base32 is considered unwanted or
not.

Anyway, please disregard the decoder for now, as it is not done. The encoder,
however passes the tests, and I don't think it deviates too far from the base64
implementation.

It supports two standard alphabet sets, the usual, and a hex alphabet - both as
described in the RFC 4648[1].

What do you think? Could this find a spot alongside base64? If so, I'll get
started with the decoder.

Looking forward to further discussion!

[1]: https://datatracker.ietf.org/doc/html/rfc4648