fetch: Disallow `--save` when fetching by file path
The purpose of `zig fetch` is to copy a package to the global cache to
avoid needing to access the network. `path` entries in a build.zig.zon
are for relative paths inside a package. It does not make sense to
`--save` a package fetched by a (CWD-relative) file path.